adding role based crud for products

+++ app/api/products/[id]/route.ts
@@ -3,6 +3,7 @@ import {
     editProduct,
     getSingleProduct,
 } from '@/app/api/products/controller'
+import { requireAdmin } from '@/utils/auth-utils'
 import { NextRequest } from 'next/server'
 
 export async function GET(
@@ -15,6 +16,10 @@ export async function GET(
 }
 
 export async function PUT(request: NextRequest) {
+    const adminResponse = await requireAdmin(request)
+    if (adminResponse) {
+        return adminResponse
+    }
     return await editProduct(request)
 }
 
@@ -22,6 +27,10 @@ export async function DELETE(
     request: NextRequest,
     { params }: { params: Promise<{ id: string }> }
 ) {
+    const adminResponse = await requireAdmin(request)
+    if (adminResponse) {
+        return adminResponse
+    }
     const id = (await params).id
     return await deleteProduct(Number(id))
 }
